# docs/swagger.yaml
definitions:
  model.Product:
    properties:
      country:
        type: string
      description:
        type: string
      discount:
        description: MaxDiscountPercent
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        description: UnitPrice
        type: number
      quantity:
        type: integer
      region:
        type: string
    type: object
info:
  contact: {}
  description: This is Product Info System Server for Central Warehouse.
  title: Product Info System API
  version: "1.0"
paths:
  /product:
    post:
      consumes:
      - application/json
      description: create method for adding new product info product information system
        on the server
      parameters:
      - description: Product to add Product information system
        in: body
        name: addProduct
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: add a new Product into product information system
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete method for particular product's information on the server
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Delete Product from product information system
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get method for particular product information from server
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      summary: Get particular Product's information from the server
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Update method for particular product's information on the server
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product details to update
        in: body
        name: updateProduct
        required: true
        schema:
          $ref: '#/definitions/model.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update Product's information
      tags:
      - Product
  /products:
    get:
      consumes:
      - application/json
      description: Get method for listing all the products from server
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      summary: List all Products information from the server
      tags:
      - Products
swagger: "2.0"